refactor(tcp): split connection concerns into modules
Move MBAP parsing, timeout config, pending transaction tracking,
connected-state reader lifecycle, and cancellation-safe writes out of
the public connection handle so future TCP client and server work can
build on stable concern-focused seams without changing behavior.
